Renamed module for new driver + moved program1 in it's entirety into it
